+++ BitmapFont.js
@@ -1306,4 +1306,4 @@ export class BitmapFont {
 }
 export default BitmapFont;
 
-export const VERSION = '1.7.0';
+export const VERSION = '1.8.0';